+++ external/autotest/site_utils/job_directories.py
@@ -2,6 +2,7 @@ import abc
 import datetime
 import glob
 import json
+import logging
 import os
 import re
 import shutil
@@ -12,14 +13,13 @@ from autotest_lib.client.common_lib import utils
 from autotest_lib.server.cros.dynamic_suite import constants
 from autotest_lib.server.cros.dynamic_suite import frontend_wrappers
 
+try:
+    from chromite.lib import metrics
+except ImportError:
+    metrics = utils.metrics_mock
 
-_AFE = frontend_wrappers.RetryingAFE()
 
 SPECIAL_TASK_PATTERN = '.*/hosts/[^/]+/(\d+)-[^/]+'
-JOB_PATTERN = '.*/(\d+)-[^/]+'
-# Pattern of a job folder, e.g., 123-debug_user, where 123 is job id and
-# debug_user is the name of user starts the job.
-JOB_FOLDER_PATTERN = '.*/(\d+-[^/]+)'
 
 def is_job_expired(age_limit, timestamp):
   """Check whether a job timestamp is older than an age limit.
@@ -48,8 +48,8 @@ def get_job_id_or_task_id(result_dir):
             For special task:
             /usr/local/autotest/results/hosts/chromeos1-rack5-host6/1343-cleanup
 
-    @returns: integer representing the job id or task id. Returns None if fail
-              to parse job or task id from the result_dir.
+    @returns: str representing the job id or task id. Returns None if fail
+        to parse job or task id from the result_dir.
     """
     if not result_dir:
         return
@@ -59,35 +59,34 @@ def get_job_id_or_task_id(result_dir):
     # Try to get the job ID from the last pattern of number-text. This avoids
     # issue with path like 123-results/456-debug_user, in which 456 is the real
     # job ID.
-    m_job = re.findall(JOB_PATTERN, result_dir)
-    if m_job:
-        return int(m_job[-1])
-    m_special_task = re.match(SPECIAL_TASK_PATTERN, result_dir)
-    if m_special_task:
-        return int(m_special_task.group(1))
-    m_ssp_job_pattern = re.match(ssp_job_pattern, result_dir)
-    if m_ssp_job_pattern and utils.is_in_container():
-        return int(m_ssp_job_pattern.group(1))
-
-
-def get_job_folder_name(result_dir):
-    """Extract folder name of a job from result_dir.
-
-    @param result_dir: path to the result dir.
-            For test job:
-            /usr/local/autotest/results/2032-chromeos-test/chromeos1-rack5-host6
-            The hostname at the end is optional.
-            For special task:
-            /usr/local/autotest/results/hosts/chromeos1-rack5-host6/1343-cleanup
-
-    @returns: The name of the folder of a job. Returns None if fail to parse
-            the name matching pattern JOB_FOLDER_PATTERN from the result_dir.
-    """
-    if not result_dir:
-        return
-    m_job = re.findall(JOB_FOLDER_PATTERN, result_dir)
+    m_job = re.findall('.*/(\d+)-[^/]+', result_dir)
     if m_job:
         return m_job[-1]
+    m_special_task = re.match(SPECIAL_TASK_PATTERN, result_dir)
+    if m_special_task:
+        return m_special_task.group(1)
+    m_ssp_job_pattern = re.match(ssp_job_pattern, result_dir)
+    if m_ssp_job_pattern and utils.is_in_container():
+        return m_ssp_job_pattern.group(1)
+    return _get_swarming_run_id(result_dir)
+
+
+def _get_swarming_run_id(path):
+    """Extract the Swarming run_id for a Skylab task from the result path."""
+    # Legacy swarming results are in directories like
+    #   .../results/swarming-3e4391423c3a4311
+    # In particular, the ending digit is never 0
+    m_legacy_path = re.match('.*/swarming-([0-9a-fA-F]*[1-9a-fA-F])$', path)
+    if m_legacy_path:
+        return m_legacy_path.group(1)
+    # New style swarming results are in directories like
+    #   .../results/swarming-3e4391423c3a4310/1
+    # - Results are one directory deeper.
+    # - Ending digit of first directory is always 0.
+    m_path = re.match('.*/swarming-([0-9a-fA-F]*)0/([1-9a-fA-F])$', path)
+    if m_path:
+      return m_path.group(1) + m_path.group(2)
+    return None
 
 
 class _JobDirectory(object):
@@ -178,11 +177,10 @@ class RegularJobDirectory(_JobDirectory):
         _remove_log_directory_contents(dirname)
 
     # Finally check if there's anything left to offload.
-    if not os.listdir(self.dirname):
+    if os.path.exists(self.dirname) and not os.listdir(self.dirname):
       shutil.rmtree(self.dirname)
       return False
     return True
-
 
   def get_timestamp_if_finished(self):
     """Get the timestamp to use for finished jobs.
@@ -190,11 +188,11 @@ class RegularJobDirectory(_JobDirectory):
     @returns the latest hqe finished_on time. If the finished_on times are null
              returns the job's created_on time.
     """
-    entry = _AFE.get_jobs(id=self._id, finished=True)
+    entry = _cached_afe().get_jobs(id=self._id, finished=True)
     if not entry:
       return None
-    hqes = _AFE.get_host_queue_entries(finished_on__isnull=False,
-                                       job_id=self._id)
+    hqes = _cached_afe().get_host_queue_entries(finished_on__isnull=False,
+                                                job_id=self._id)
     if not hqes:
       return entry[0].created_on
     # While most Jobs have 1 HQE, some can have multiple, so check them all.
@@ -224,5 +222,66 @@ class SpecialJobDirectory(_JobDirectory):
     super(SpecialJobDirectory, self).__init__(resultsdir)
 
   def get_timestamp_if_finished(self):
-    entry = _AFE.get_special_tasks(id=self._id, is_complete=True)
+    entry = _cached_afe().get_special_tasks(id=self._id, is_complete=True)
     return entry[0].time_finished if entry else None
+
+
+_OFFLOAD_MARKER = ".ready_for_offload"
+_marker_parse_error_metric = metrics.Counter(
+    'chromeos/autotest/gs_offloader/offload_marker_parse_errors',
+    description='Errors parsing the offload marker file')
+
+
+class SwarmingJobDirectory(_JobDirectory):
+  """Subclass of _JobDirectory for Skylab swarming jobs."""
+
+  @classmethod
+  def get_job_directories(cls):
+    """Return a list of directories of jobs that need offloading."""
+    # Legacy swarming results are in directories like
+    #   .../results/swarming-3e4391423c3a4311
+    # In particular, the ending digit is never 0
+    jobdirs = [d for d in glob.glob('swarming-[0-9a-f]*[1-9a-f]')
+                 if os.path.isdir(d)]
+    # New style swarming results are in directories like
+    #   .../results/swarming-3e4391423c3a4310/1
+    # - Results are one directory deeper.
+    # - Ending digit of first directory is always 0.
+    new_style_topdir = [d for d in glob.glob('swarming-[0-9a-f]*0')
+                        if os.path.isdir(d)]
+    for topdir in new_style_topdir:
+      subdirs = [d for d in glob.glob('%s/[1-9a-f]*' % topdir)
+                 if os.path.isdir(d)]
+      jobdirs += subdirs
+    return jobdirs
+
+
+
+  def get_timestamp_if_finished(self):
+    """Get the timestamp to use for finished jobs.
+
+    @returns the latest hqe finished_on time. If the finished_on times are null
+             returns the job's created_on time.
+    """
+    marker_path = os.path.join(self.dirname, _OFFLOAD_MARKER)
+    try:
+      with open(marker_path) as f:
+        ts_string = f.read().strip()
+    except (OSError, IOError) as e:
+      return None
+    try:
+      ts = int(ts_string)
+      return time_utils.epoch_time_to_date_string(ts)
+    except ValueError as e:
+      logging.debug('Error parsing %s for %s: %s',
+                    _OFFLOAD_MARKER, self.dirname, e)
+      _marker_parse_error_metric.increment()
+      return None
+
+
+_AFE = None
+def _cached_afe():
+  global _AFE
+  if _AFE is None:
+    _AFE = frontend_wrappers.RetryingAFE()
+  return _AFE
